seccion, noticia y contenido fix
seccion funcionalidad de archivos

noticia y contenido manejo de errores con response entity

+++ src/main/java/co/edu/unicauca/APIHappLab/repository/I_seccion_repository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import co.edu.unicauca.APIHappLab.model.seccion;
 
 @Repository
-public interface I_seccion_repository extends MongoRepository<seccion,Double>{
+public interface I_seccion_repository extends MongoRepository<seccion,Integer>{
 }
